I think this is everything
i tried to make sure that all of the requirements were added.  I think
we still need to write a lab report... though what we are doing to
complete the requirements should be obvious from the website.  I put
Section 2 in the requiremnts before section 1 because it makes more
sense for it to be there.